error hadling and relate user to diary

+++ app/Http/Controllers/DiaryController.php
@@ -5,12 +5,14 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Diary;
 use App\Http\Requests\CreateDiary;
+use Illuminate\Support\Facades\Auth;
 
 class DiaryController extends Controller
 {
     public function index()
     {
-        $diaries = Diary::all(); //diariesテーブルのデータを全件取得
+
+        $diaries = Diary::all();
 
         return view('diaries.index',[
             'diaries' => $diaries,
@@ -23,28 +25,33 @@ class DiaryController extends Controller
     }
 
     public function store(CreateDiary $request)
-    {
+    {        
         $diary = new Diary(); //Diaryモデルをインスタンス化
 
         $diary->title = $request->title; //画面で入力されたタイトルを代入
         $diary->body = $request->body; //画面で入力された本文を代入
+        $diary->user_id = Auth::user()->id;
         $diary->save(); //DBに保存
 
         return redirect()->route('diary.index'); //一覧ページにリダイレクト
     }
 
-    public function edit(int $id)
+    public function edit(Diary $diary)
     {
-        $diary = Diary::find($id);
+        if (Auth::user()->id !== $diary->user_id) {
+            abort(403);
+        }
 
         return view('diaries.edit', [
             'diary' => $diary,
         ]);
     }
 
-    public function update(int $id, CreateDiary $request)
+    public function update(Diary $diary, CreateDiary $request)
     {
-        $diary = Diary::find($id);
+        if (Auth::user()->id !== $diary->user_id) {
+            abort(403);
+        }
 
         $diary->title = $request->title; //画面で入力されたタイトルを代入
         $diary->body = $request->body; //画面で入力された本文を代入
@@ -53,9 +60,12 @@ class DiaryController extends Controller
         return redirect()->route('diary.index'); //一覧ページにリダイレクト
     }
 
-    public function destroy(int $id)
+    public function destroy(Diary $diary)
     {
-        $diary = Diary::find($id);
+        if (Auth::user()->id !== $diary->user_id) {
+            abort(403);
+        }
+        
         $diary->delete();
 
         return redirect()->route('diary.index');
